+++ AkhshamBazari/Controllers/Base/ControllerBase.cs
@@ -1,0 +1,6 @@
+namespace AkhshamBazari.Controllers.Base;
+
+public class ControllerBase
+{
+    
+}
